extract Actions from RunOnceHandler
This introduces an ActionRunner object which abstracts the sequential
execution of Actions, and enables you to cancel them. It performs cleanup on
the completed actions in reverse order.

This is to enable controlled shutdown/short-circuiting of in-flight RunOnces.
It will later be used for ExecuteAction's own sequence of actions.

There is a hole in RunOnceHandler's tests introduced, which will be fixed in
later commits. (The glue code is untested.)

[#66631332]

Signed-off-by: Abhijit Hiremagalur <abhi@pivotallabs.com>