initial monitoring implementation
* container.Action no longer a pointer
* added Serial step
* refactored lots of things
* switch executor integration to using real garden
* refactored steps

an executor container now specifies three different actions:

  - setup: executed before starting
  - action: the primary action to execute
  - monitor: an action executed periodically after setup, while the
    primary action is running

if the monitor action succeeds and then fails, the container is stopped
and its health is reported as 'down'.

if the action fails, the container is stopped.

if the action succeeds, but monitoring is still successful, the health
of the container is still 'up' (this allows processes to daemonize).

if the monitor action is not provided, the health is 'unmonitored', and
the container completes as soon as the primary action exits.

also, the integration tests now run against a real Garden server.

Signed-off-by: Ted Young <tyoung@pivotallabs.com>
Signed-off-by: Matthew Sykes <matthew.sykes@gmail.com>