if told to stop, don't complete with failure
this fixes the case where scaling down causes us to continously bring it
back (treating it as a crash).

also replaces the previous logic ensuring we cancel monitor before the
action. previously this was done to prevent the monitor from failing (if
it saw the run action die), which is no longer a concern with this
approach.

caveat: canceled tasks will appear successful; we'll see if this becomes
a problem in the future. the fact that the caller requested it to be
canceled means they should already know the result anyway.

[#83638710]

Signed-off-by: Georgi Sabev <georgethebeatle@gmail.com>