Refactor NewTransformer to use functional Option argument
Adding new arguments to many of our initializers is annoying because you
have to change lots of different places in order to make it compile.
Changing any functionality that is optional to use this new functional
style enables easy integration of new features and allows logical
grouping of multiple values.

[#153057330]

Signed-off-by: Brandon Shroyer <bshroyer@pivotal.io>
Signed-off-by: Chris Piraino <cpiraino@pivotal.io>