Use http protocol for fake submodule repo
Newer versions of git are disallowing the file scheme for git repos without special flags. Rather than try to untagle the mixture of http:// and file:// schemes in these tests, we simply use http:// everywhere.

This does, however, require that the local HTTP server that serves the mock git repos be started in the SynchronizedBeforeSuite, so that 'git submodule add ...' can fetch the submodule via http