Use natively built executable name instead of hardcoded 'hello' string
These tests were failing on Windows when the test executable was explicitly
installed as 'hello', as opposed to using the native go build output of 'hello.exe'.

Changing all invocations to use the 'basename' of the go built executable (in the case of
windows an '.exe' extension is added) allowed the failing tests to succeed.

Signed-off-by: David Sabeti <sabetid@vmware.com>