Update garden container logging tags
When apps are renamed the logging tags are updated such that the new
name is reflected in the log output.

This change updates the tags in garden with that updated information. A
new component will be reading these updated tags from garden and
updating the networking config so that the dynamic updates of security
groups include the correct app name in the logs.

Signed-off-by: Matthew Kocher <mkocher@vmware.com>
Signed-off-by: Rebecca Roberts <robertsre@vmware.com>